Bugfix: 1. Get the operand with right index for printBitRepeat.         2. Fix the argument type of VASTUse constructor that construct an            immediate use.

+++ include/vtm/VerilogAST.h
@@ -126,7 +126,7 @@ public:
     UseKind = USE_Value;
   }
 
-  VASTUse(unsigned immVal, uint8_t width) : UB(width), LB(0) {
+  VASTUse(int64_t immVal, uint8_t width) : UB(width), LB(0) {
     Data.ImmVal = immVal;
     UseKind = USE_Immediate;
   }
